Added Fragment support for Course/Assignment connection So far, it's just a basic framework. Basically, when you click on a course, it displays all the assignments with it.  For this, it just generates dummy data for every course.  When we finally set Parse up with pulling from the database, it'll dynamically add to the ListView based upon the course being selected.
My thought was to have a dialog display when an assignment is selected, thus allowing the user to edit anything about it.  As of right now, the assignment only shows it's name, but it'll also display pertinent information like when it's due, and the priority.  When the user clicks on the assignment, it'll display the aforementioned dialog that will then show more, as well as allow the user to edit anything about the assignment.

For right now, the fragment setup is VERY basic, but works on both handheld and tablet devices the way they are supposed to (I tested the app on both a Nexus 4 and a Nexus 7 running 4.2.2).  When an assignment is clicked, a toast pops up saying that the functionality here isn't available yet, but will be by final submission.